Extract taskpane inventory / Content formatting helpers
Move 10 pure presentation helpers from taskpane.js into the new
taskpane-inventory-formatters.js module. No behavior changes.
